functions/pbp-parse/requirements.txt
functions-framework==3.*
duckdb==1.3.2
google-cloud-secret-manager
pandas